+++ src/markdown/parse_image.js
@@ -1,4 +1,6 @@
 /** @module */
+const escapeStyle = target =>
+  target.replace(/[\\;:()]/g, matched => `\\${matched[0]}`)
 
 /**
  * Marpit parse image plugin.
@@ -9,8 +11,77 @@
  *
  * @alias module:markdown/parse_image
  * @param {MarkdownIt} md markdown-it instance.
+ * @param {Object} [opts]
+ * @param {boolean} [opts.filters=true] Switch feature to support CSS filters.
  */
-function parseImage(md) {
+function parseImage(md, opts = {}) {
+  const pluginOptions = { filters: true, ...opts }
+  const optionMatchers = new Map()
+
+  // The scale percentage for resize
+  // TODO: Implement cross-browser image zoom without affecting DOM tree
+  // (Pre-released Marp uses `zoom` but it has not supported in Firefox)
+  optionMatchers.set(/^(\d*\.)?\d+%$/, matches => ({ size: matches[0] }))
+
+  if (pluginOptions.filters) {
+    // CSS filters
+    optionMatchers.set(/^blur(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['blur', escapeStyle(matches[1] || '10px')]],
+    }))
+    optionMatchers.set(/^brightness(?::(.+))?$/, (matches, meta) => ({
+      filters: [
+        ...meta.filters,
+        ['brightness', escapeStyle(matches[1] || '1.5')],
+      ],
+    }))
+    optionMatchers.set(/^contrast(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['contrast', escapeStyle(matches[1] || '2')]],
+    }))
+    optionMatchers.set(
+      /^drop-shadow(?::(.+?),(.+?)(?:,(.+?))?(?:,(.+?))?)?$/,
+      (matches, meta) => {
+        const args = matches
+          .slice(1)
+          .filter(v => v)
+          .map(arg => {
+            const colorFunc = arg.match(/^(rgba?|hsla?)\((.*)\)$/)
+
+            return colorFunc
+              ? `${colorFunc[1]}(${escapeStyle(colorFunc[2])})`
+              : escapeStyle(arg)
+          })
+
+        return {
+          filters: [
+            ...meta.filters,
+            ['drop-shadow', args.join(' ') || '0 5px 10px rgba(0,0,0,.4)'],
+          ],
+        }
+      }
+    )
+    optionMatchers.set(/^grayscale(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['grayscale', escapeStyle(matches[1] || '1')]],
+    }))
+    optionMatchers.set(/^hue-rotate(?::(.+))?$/, (matches, meta) => ({
+      filters: [
+        ...meta.filters,
+        ['hue-rotate', escapeStyle(matches[1] || '180deg')],
+      ],
+    }))
+    optionMatchers.set(/^invert(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['invert', escapeStyle(matches[1] || '1')]],
+    }))
+    optionMatchers.set(/^opacity(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['opacity', escapeStyle(matches[1] || '.5')]],
+    }))
+    optionMatchers.set(/^saturate(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['saturate', escapeStyle(matches[1] || '2')]],
+    }))
+    optionMatchers.set(/^sepia(?::(.+))?$/, (matches, meta) => ({
+      filters: [...meta.filters, ['sepia', escapeStyle(matches[1] || '1')]],
+    }))
+  }
+
   md.inline.ruler2.push('marpit_parse_image', ({ tokens }) => {
     tokens.forEach(token => {
       if (token.type === 'image') {
@@ -24,10 +95,28 @@ function parseImage(md) {
         }
 
         options.forEach(opt => {
-          // TODO: Implement cross-browser image zoom without affecting DOM tree
-          // (Pre-released Marp uses `zoom` but it has not supported in Firefox)
-          if (opt.match(/^(\d*\.)?\d+%$/)) token.meta.marpitImage.size = opt
+          optionMatchers.forEach((mergeFunc, regexp) => {
+            const matched = opt.match(regexp)
+
+            if (matched)
+              token.meta.marpitImage = {
+                ...token.meta.marpitImage,
+                ...mergeFunc(matched, {
+                  filters: [],
+                  ...token.meta.marpitImage,
+                }),
+              }
+          })
         })
+
+        // Build and apply filter style
+        if (token.meta.marpitImage.filters) {
+          token.meta.marpitImage.filter = token.meta.marpitImage.filters
+            .reduce((arr, fltrs) => [...arr, `${fltrs[0]}(${fltrs[1]})`], [])
+            .join(' ')
+
+          token.attrJoin('style', `filter:${token.meta.marpitImage.filter};`)
+        }
       }
     })
   })